lp-1981 feedback changes
